refactor(generator): 优化文件合并服务的代码结构
- 添加自动生成代码区域的开始和结束标记注释
- 引入默认初始容量常量，替换魔法数字
- 调整自定义区块提取方法中集合初始化容量为默认常量
- 枚举DiffLineType中增加未变更、新增、删除行的注释说明
- 优化TokenCostCalculator中字符比例计算，使用BigDecimal并转换为double进行计算

+++ ydsz-agent/ydsz-agent-server/src/main/java/com/njydsz/agent/server/chat/TokenCostCalculator.java
@@ -1,5 +1,6 @@
 package com.njydsz.agent.server.chat;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.stereotype.Component;
@@ -74,8 +75,8 @@ public class TokenCostCalculator {
         }
       }
     }
-    double charRatio = properties.getMemory().getTokenCharRatio();
-    int estimatedPromptTokens = Math.max(1, (int) Math.ceil(totalChars / charRatio));
+    BigDecimal charRatio = properties.getMemory().getTokenCharRatio();
+    int estimatedPromptTokens = Math.max(1, (int) Math.ceil(totalChars / charRatio.doubleValue()));
     double unitPrice = resolveUnitPrice(request.getModel());
     return CostEstimate.estimate(
         estimatedPromptTokens, request.getMaxTokens(), request.getModel(), unitPrice);
